starrwork/updated profile pages, reorganized components
