+++ src/components/RoutesAll.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import { Register } from './usercomponents/Register';
-import { Profile } from './usercomponents/Profile';
-import { ProfileEdit } from './usercomponents/ProfileEdit';
+import { Profile } from './profilecomponents/Profile';
+import { ProfileEdit } from './profilecomponents/ProfileEdit';
 import { Wins } from './wincomponents/Wins';
 import { WinForm } from './wincomponents/WinForm';
 import { WinDetail } from './wincomponents/WinDetail';
